Renamed Quant to IntQuant, preserved backwards compatibility for refs to quant.py and Quant.
