+++ src/qonnx/custom_op/general/__init__.py
@@ -31,9 +31,9 @@ from qonnx.custom_op.general.debugmarker import DebugMarker
 from qonnx.custom_op.general.floatquant import FloatQuant
 from qonnx.custom_op.general.genericpartition import GenericPartition
 from qonnx.custom_op.general.im2col import Im2Col
+from qonnx.custom_op.general.intquant import IntQuant
 from qonnx.custom_op.general.maxpoolnhwc import MaxPoolNHWC
 from qonnx.custom_op.general.multithreshold import MultiThreshold
-from qonnx.custom_op.general.quant import Quant
 from qonnx.custom_op.general.quantavgpool2d import QuantAvgPool2d
 from qonnx.custom_op.general.trunc import Trunc
 from qonnx.custom_op.general.xnorpopcount import XnorPopcountMatMul
@@ -47,7 +47,8 @@ custom_op["GenericPartition"] = GenericPartition
 custom_op["MultiThreshold"] = MultiThreshold
 custom_op["XnorPopcountMatMul"] = XnorPopcountMatMul
 custom_op["Im2Col"] = Im2Col
-custom_op["Quant"] = Quant
+custom_op["IntQuant"] = IntQuant
+custom_op["Quant"] = IntQuant
 custom_op["Trunc"] = Trunc
 custom_op["BipolarQuant"] = BipolarQuant
 custom_op["FloatQuant"] = FloatQuant
